Done Good Settings

+++ database/migrations/2024_11_13_162201_create_doctors_table.php
@@ -13,14 +13,14 @@ return new class extends Migration
     {
         Schema::create('doctors', function (Blueprint $table) {
             $table->id();
-            $table->bigInteger('doctor_code')->unique();
+            $table->integer('doctor_code')->unique();
             $table->string('id_number',14);
             $table->string('name',50);
             $table->string('mobile',20);
             $table->string('address',300);
             $table->string('title',300);
             $table->string('email',300)->unique();
-            $table->integer('gender')->comment('1:Male,2:Female');
+            $table->tinyInteger('gender')->comment('1:Male,2:Female');
             $table->integer('status')->default(1)->nullable();
             $table->foreignId('nationality_id')->references('id')->on('nationalities')->onUpdate('cascade');
             $table->foreignId('specialization_id')->references('id')->on('specializations')->onUpdate('cascade');
